Public

ImagPhotonics.ONEFORTH — Constant

A constant representing one-fourth (0.25).

ImagPhotonics.ONEHALF — Constant

A constant representing one-half (0.5).

ImagPhotonics.AbstractComponent — Type

abstract type AbstractComponent{T<:Unitful.Quantity} end

Represents a generic photonic component type with dimensions and units specified in the type signature. The dimensions and unit cannot be an empty set.

The type parameter T<:Unitful.Quantity. The layer type is used for representation in the GDSII files, and the datatype is used to define the material parameters.

See also Waveguide, Taper, Device, IIDD

ImagPhotonics.AbstractConnectorFunction — Type

abstract type AbstractConnectorFunction<:AbstractIsoparametricFunction end

Represents a generic connector function type, which is a subtype of AbstractIsoparametricFunction. It is used to define the connection between photonic components.

ImagPhotonics.AbstractIsoparametricFunction — Type

abstract type AbstractIsoparametricFunction end

Represents a generic isoparametric function type, with the order defined in the type signature. The order is allowed to be an empty set, in which case a default linear isoparametric function is returned.

ImagPhotonics.Connector_Type — Type

An enumeration representing the connector type for a photonic device.

C_INPUT: Indicates the port is an input port.

C_OUTPUT: Indicates the port is an output port.

ImagPhotonics.Device — Type

mutable struct Device

A mutable struct representing a 2D photonic device. It inherits from AbstractComponent. It generated the photonic structure used in the optimization process.

The type parameter T<:Unitful.Quantity.

Fields

width::T : Width of the device.

length::T : Length of the device.

curvature_left::T : Curvature on the left side of the device.

curvature_right::T : Curvature on the right side of the device.

curvature_up::T : Curvature on the top side of the device.

curvature_down::T : Curvature on the bottom side of the device.

number_waveguides::NTuple{2,Union{Int,Vector{Int}}} : Tuple containing the number of input and output waveguides.

location::NTuple{2,Union{LocationType,Vector{LocationType}}} : Tuple containing the location of input and output waveguides.

points::Array{T, 2} : The array representing the device points.

pin::Vector{Port} : The vector representing the input ports of the device.

pout::Vector{Port} : The vector representing the output ports of the device.

shape_function::ShapeFunction2D : The shape function used for the device.

layer::Int : The layer for the device.

datatype::Int : The datatype for the device.

Constructor

Device(length, width; number_waveguides=(4,4),spacing=(5.0u"μm",5.0u"μm"),
location=(Location_Type(0),Location_Type(1)), 
curvature_left=0.0u"μm", curvature_right=0.0u"μm", 
curvature_up  =0.0u"μm", curvature_down =0.0u"μm",
layer=0, datatype=1)

Generates a device with number_waveguides input and output for gdsii files and photonic computations.

The length and width are provided by the user. The waveguides are located at location and are separated by spacing. The first and second entry of number_waveguides,location, and spacing represent the input and output respectively. The curvature parameters curvature_down,curvature_up,curvature_left,curvature_right represent the shift from the linearity in the middle of each respective edge, if all 4 parameters are 0, the device is a rectangle.

The layer type is used for representation in the gdsii files and the datatype is used to define the material parameters.

Arguments

length              :: T
width               :: T 
number_waveguides   :: NTuple{2,Union{Int,Vector{Int}}} 
spacing             :: NTuple{2,Union{T,Vector{T}}}         
location            :: NTuple{2,Union{Location_Type,Vector{Location_Type}}}         
curvature_left      :: T 
curvature_right     :: T 
curvature_up        :: T 
curvature_down      :: T
layer               :: Int
datatype            :: Int

Examples

julia> d=Device(10.0u"μm",20.0u"μm")
Photonic Device
---------------
length: 10.0 μm
width: 20.0 μm
pᵢₙ  : left-4
pₒᵤₜ  : right-4
curvature left: 0.0 μm
curvature right: 0.0 μm
curvature up: 0.0 μm
curvature down: 0.0 μm
layer: 0
material: 1
Shape functions: N₁, N₂, N₃, N₄

ImagPhotonics.Location_Type — Type

An enumeration representing the location type of a port on a photonic device.

LOC_LEFT: Indicates the port is located on the left side of the device.

LOC_RIGHT: Indicates the port is located on the right side of the device.

LOC_UP: Indicates the port is located on the upper side of the device.

LOC_DOWN: Indicates the port is located on the lower side of the device.

ImagPhotonics.Monitor_Type — Type

An enumeration representing the monitor type for a photonic device.

GENERIC: Indicates a generic monitor.

OPTIMIZATION: Indicates an optimization monitor.

SOURCE: Indicates a source monitor.

ImagPhotonics.PIC_IIDD — Type

...

ImagPhotonics.PhotonicComponent_Type — Type

An enumeration representing various photonic component types.

DEVICE: Indicates a generic photonic device.

WAVEGUIDE: Indicates a waveguide component.

TAPER: Indicates a taper component.

TERMINATION: Indicates a termination component.

RING: Indicates a ring resonator component.

ImagPhotonics.Port — Type

struct Port

Represents a port location on an AbstractComponent in element units ξ and η.

Fields:

x: Real number representing the x-coordinate of the port location.

y: Real number representing the y-coordinate of the port location.

ImagPhotonics.ShapeFunction2D — Method

(sf::ShapeFunction2D)(ξ, η, points)

Returns the value in the xy space, given the element coordinates ξ, η and the control points.

EXAMPLE

julia> points[0. 0.; 1. 0.; 1. 1.; 0. 1.]
4×2 Matrix{Float64}:
 0.0  0.0
 1.0  0.0
 1.0  1.0
 0.0  1.0
 julia> sf=ShapeFunction2D(0,0,points);
 (0.5, 0.5)

ImagPhotonics.Taper — Type

Taper{T} <: AbstractComponent{T}

A concrete subtype of AbstractComponent representing a taper or termination for a waveguide in photonic devices.

Fields

width_initial : Initial width of the taper

width_final : Final width of the taper

length : Length of the taper

m : Design parameter

points : Array of points representing the taper

pin : Input port

pout : Output port

shape_function : TaperFunction object representing the shape of the taper

layer : Layer of the device

datatype : Datatype of the device

Constructors

Taper(;width_initial, width_final, length, m, layer, ComponentType, datatype, kwargs...)

Arguments

width_initial::T=0.5u"μm" : Initial width of the taper

width_final::T=1.0u"μm" : Final width of the taper

length::T=1.0u"μm" : Length of the taper

m::Real=0.0 : Design parameter

layer::Int=0 : Layer of the device

ComponentType::PhotonicComponent_Type=TAPER : Type of the component (Taper or Termination)

datatype::Int=0 : Datatype of the device

kwargs... : Additional keyword arguments for the TaperFunction constructor

Examples

julia> Taper()

Taper
---------------
length: 1.0 μm
initial width: 0.5 μm
final width: 1.0 μm
pᵢₙ  : ImagPhotonics.Port(-1.0, 0.0)
pᵢₙ  : ImagPhotonics.Port(-1.0, 0.0)
pᵢₙ  : ImagPhotonics.Port(-1.0, 0.0)
pᵢₙ  : ImagPhotonics.Port(-1.0, 0.0)
pₒᵤₜ : ImagPhotonics.Port(-1.0, 0.0)
layer: 0
material: 268136720

julia> Taper(ComponentType=TERMINATION)

[ Info: width_final=0.0 μm, the structure is termination
Termination
---------------
length: 1.0 μm
width: 0.5 μm
pᵢₙ  : ImagPhotonics.Port(-1.0, 0.0)
pᵢₙ  : ImagPhotonics.Port(-1.0, 0.0)
pᵢₙ  : ImagPhotonics.Port(-1.0, 0.0)
pᵢₙ  : ImagPhotonics.Port(-1.0, 0.0)
pₒᵤₜ : ImagPhotonics.Port(-1.0, 0.0)
layer: 0
material: 268158704

ImagPhotonics.TaperFunction — Type

TaperFunction

A concrete subtype of ImagPhotonics.AbstractConnectorFunction, representing a 2D shape function. ShapeFunction2D generates the basis functions Na, and its derivatives Naξ and Naη with respect to the element space ξ and η.A mutable struct representing a taper function for a photonic device.

Fields

Na : Shape function

Na_ξ : Tuple of functions representing the partial derivatives of Na with respect to ξ

Na_η : Tuple of functions representing the partial derivatives of Na with respect to η

Affinity : Affinity matrix used for affine transformations

_rotation : Rotation angle

_translation : Tuple representing translation along x and y

_scale : Tuple representing scaling factors along x and y

_mirror_x : Bool for mirroring along the x-axis

_mirror_y : Bool for mirroring along the y-axis

_operation_order : Array of transformation types representing the order of operations

Constructor

TaperFunction(L::Number, w₁::Number, w₂::Number; m = 1.0,

rotation::Real=0, translation::Tuple{Real,Real}=(0.,0.), scale::Tuple{Real,Real}=(1.,1.), mirrorx::Bool=false, mirrory::Bool=false, operationorder=[AFFINETRANSLATION,AFFINEROTATION,AFFINESCALE,AFFINEMIRRORX,AFFINEMIRRORY])

Create a taper function, which is a parametric function that represents a tapered shape (e.g., a trapezoid) with linearly changing width along its length.

Arguments

L: The length of the tapered shape.

w₁: The width at the starting point (x = 0) of the tapered shape.

w₂: The width at the end point (x = L) of the tapered shape.

m: A power factor affecting the tapering shape (default: 1.0).

rotation: Rotation angle in radians (default: 0).

translation: Translation vector as a tuple (x, y) (default: (0.,0.)).

scale: Scaling factor, either a single value or a tuple (xscale, yscale) (default: (1.,1.)).

mirror_x: Reflection along the x-axis (true/false) (default: false).

mirror_y: Reflection along the y-axis (true/false) (default: false).

operation_order: An array specifying the order of transformations (default: [AFFINETRANSLATION, AFFINEROTATION, AFFINESCALE, AFFINEMIRRORX, AFFINEMIRROR_Y]).

Returns

A TaperFunction object representing the tapered shape with the specified properties.

ImagPhotonics.Transformation_Type — Type

An enumeration representing the transformation type for a photonic device.

AFFINE_TRANSLATION: Indicates an affine translation transformation.

AFFINE_ROTATION: Indicates an affine rotation transformation.

AFFINE_SCALE: Indicates an affine scale transformation.

AFFINE_MIRROR_X: Indicates an affine mirror transformation along the X axis.

AFFINE_MIRROR_Y: Indicates an affine mirror transformation along the Y axis.

ImagPhotonics.Waveguide — Type

...

RefractiveIndex.RefractiveMaterial — Method

RefractiveIndex.RefractiveMaterial(name::String, dispersion::RefractiveIndex.DispersionFormula; λrange::Union{T,Tuple{T,T}} = 1.55u"μm", reference::String = "", comment::String = "", specs::Dict{Symbol, Any} = Dict{Symbol, Any}()) where T <: Union{Float64,Unitful.Quantity}

Creates a RefractiveMaterial instance with the given name, dispersion formula, wavelength range, reference, comment, and additional specifications.

#Arguments:

-name::String : Name of the refractive material. -dispersion::RefractiveIndex.DispersionFormula : Dispersion formula of the refractive material. -λrange::Union{T,Tuple{T,T}} : Wavelength range of the refractive material. If a single value is provided, it is assumed to be the upper limit, and the lower limit is set to 0. Defaults to 1.55u"μm. -reference::String : Reference information about the refractive material. Defaults to an empty string. -comment::String : Additional comments about the refractive material. Defaults to an empty string. -specs::Dict{Symbol, Any} : Additional specifications for the refractive material. Defaults to an empty dictionary. Returns:

A RefractiveIndex.RefractiveMaterial instance with the specified arguments.

ImagPhotonics.add_custom_material — Method

add_custom_material(key::String, material::RefractiveMaterial; custom_lib::CustomLib=CUSTOM_LIB) -> Nothing

Adds a custom material to the custom library.

Arguments:

-key::String: Key to associate with the custom material. -material::RefractiveMaterial: Custom material to be added to the library. -custom_lib::CustomLib: Custom library to add the material to. Defaults to CUSTOM_LIB.

Returns:

nothing.

ImagPhotonics.add_marker! — Method

function addmarker!(cell, point::Tuple{T,T}; sizemarker::T=0.1) where T <: Real

Add a square marker with a specified size to the given GDSII cell at a specified point. The marker is placed on layer 999.

Arguments:

cell: GDSII cell to which the marker will be added.

point: Tuple of real numbers representing the x and y coordinates of the center point of the marker.

size_marker: Optional real number specifying the side length of the square marker. Default is 0.1.

Returns:

The modified GDSII cell with the added marker.

ImagPhotonics.affinity_matrix — Method

affinity_matrix(;rotation, translation, scale, mirror_x, mirror_y, operation_order)

Generates an affinity matrix used for affine transformations.

Arguments

rotation::Real=0 : Rotation angle

translation::Tuple{Real,Real}=(0,0) : Tuple representing translation along x and y

scale::Union{Real,Tuple{Real,Real}}=1 : Tuple representing scaling factors along x and y

mirror_x::Bool=false : Bool for mirroring along the x-axis

mirror_y::Bool=false : Bool for mirroring along the y-axis

operation_order::Array{Transformation_Type}=[AFFINE_TRANSLATION, AFFINE_ROTATION, AFFINE_SCALE, AFFINE_MIRROR_X, AFFINE_MIRROR_Y] : Array of transformation types representing the order of operations

ImagPhotonics.construct_device — Method

constructdevice(device::Device; Nξ::Int=100, N_η::Int=100)

Constructs a photonic device based on the given device object and returns a GDSTK cell object. The function takes the following arguments:

device: A 2D photonic device object.

N_ξ: Number of ξ points (default: 100).

N_η: Number of η points (default: 100).

ImagPhotonics.create_custom_material — Method

create_custom_material(name::String, dispersion::T; custom_lib::CustomLib=CUSTOM_LIB) -> RefractiveMaterial where T <: RefractiveIndex.DispersionFormula

Creates a custom material with the given name and dispersion formula and adds it to the custom library.

Arguments:

name::String: Name of the custom material.

dispersion::T: Dispersion formula of the custom material.

custom_lib::CustomLib: Custom library to add the material to. Defaults to CUSTOM_LIB.

Returns:

A RefractiveMaterial instance with the given name and dispersion formula.

ImagPhotonics.initialize_custom_lib — Method

initialize_custom_lib(custom_lib_path::String) -> CustomLib

Initializes a custom library.

Arguments:

custom_lib_path::String: The path to the custom library file.

Returns:

A CustomLib instance with the given path and an empty dictionary if the file does not exist, or a dictionary containing the custom materials if the file exists.

ImagPhotonics.is_squarematrix — Method

function is_squarematrix(A::Matrix)

Return true if the matrix A is square.

Examples

Copy code
julia> is_squarematrix(Matrix(undef, 3, 2))
false

julia> is_squarematrix(Matrix(undef, 3, 3))
true

ImagPhotonics.load_custom_lib — Method

loadcustomlib() -> Nothing Loads the custom library from the file.

Returns:

nothing.

ImagPhotonics.normal_to_device — Function

normal_to_device(device::Device, port_number::Int=1, connector::Connector_Type=C_INPUT)

Computes the normal vector to the surface of a 2D photonic device at the specified connection point.

Arguments:

device: A 2D photonic device object (instance of Device).

port_number: The port number to compute the normal vector for (default: 1). It should be within the range of input or output ports.

connector: The connector type, either CINPUT or COUTPUT (default: C_INPUT).

Returns:

A normal vector (as an array) to the surface at the specified connection point.

ImagPhotonics.normalize — Method

normalize(x::Unitful.AbstractQuantity) normalize(x::Matrix{Unitful.AbstractQuantity})

Return a multiplicative identity for any x of type Unitful.AbstractType.

Examples

```julia-repl julia> normalize(1u"m") 1.0 m

julia> normalize(Matrix{u"m"}(I, 2, 2)) 2×2 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(m,), 𝐋, nothing}}}: 1.0 m 0.0 m 0.0 m 1.0 m

ImagPhotonics.position_device_ports — Method

position_device_ports(length::T, width::T, spacing::T, number::Int, location::Location_Type) where T <: Unitful.Quantity

Returns an array of Port objects that represent the positions of waveguides in the square element interval ξ×η ∈ [-1,1]×[-1,1] on a 2D photonic device. The function takes the following arguments:

length: Length of the device.

width: Width of the device.

spacing: Spacing between the waveguides.

number: Number of waveguides.

location: Location of the waveguides on the device (LOCLEFT, LOCRIGHT, LOCUP, LOCDOWN).

ImagPhotonics.remove_custom_material — Function

remove_custom_material(key::String, index::Union{Int, Nothing, typeof(ALL)}=nothing; custom_lib::CustomLib=CUSTOM_LIB) -> Nothing

Removes a custom material from the custom library.

Arguments:

-key::String: Key associated with the custom material. -index::Union{Int, Nothing, typeof(ALL)}: Index of the custom material in the case of a vector of materials. If nothing, an error will be raised if trying to remove an element from a vector of RefractiveMaterials. If ALL, all materials associated with the key will be removed. Defaults to nothing. -custom_lib::CustomLib: Custom library to remove the material from. Defaults to CUSTOM_LIB.

Returns:

nothing.

ImagPhotonics.save_custom_lib — Method

save_custom_lib() -> Nothing

Saves the custom library to the file.

Returns:

nothing.

ImagPhotonics.update_affinity_matrix! — Method

update_affinity_matrix!(tf::AbstractConnectorFunction;
    rotation::Union{Real,Nothing}=nothing,
    translation::Union{Tuple{Real,Real},Nothing}=nothing,
    scale::Union{Real,Tuple{Real,Real},Nothing}=nothing,
    mirror_x::Union{Bool,Nothing}=nothing,
    mirror_y::Union{Bool,Nothing}=nothing,
    operation_order::Union{Array{Transformation_Type},Nothing}=nothing)

Update the affinity matrix of a connector function with new transformation parameters. If a parameter is not provided, the existing value will be used.

Arguments

tf: The connector function object.

rotation: Rotation angle in radians (optional).

translation: Translation vector as a tuple (x, y) (optional).

scale: Scaling factor, either a single value or a tuple (xscale, yscale) (optional).

mirror_x: Reflection along the x-axis (true/false) (optional).

mirror_y: Reflection along the y-axis (true/false) (optional).

operation_order: An array specifying the order of transformations (optional).

Returns

The updated connector function object.

ImagPhotonics.update_custom_lib — Function

update_custom_lib(custom_lib_path::Union{Nothing,String}=nothing; custom_lib::Union{Nothing,Dict{String, Union{RefractiveMaterial, Vector{RefractiveMaterial}}}}=nothing) -> Nothing

Updates a custom library.

Arguments:

-custom_lib_path::Union{Nothing,String}: Path to the custom library file. If nothing, reinitializes the DEFAULTCUSTOMLIBPATH. Defaults to nothing. customlib::Union{Nothing,Dict{String, Union{RefractiveMaterial, Vector{RefractiveMaterial}}}}: A custom library to be used. If nothing, the current library is used. Defaults to nothing.

Returns:

nothing

ImagPhotonics.update_shape_function — Method

update_shape_function(order::NTuple{4,Int}, length_device::T, width_device::T) where T

Updates the shape function of a 2D photonic device based on the given order. The function takes the following arguments:

order: A tuple containing the order of curvatures (down, left, up, right).

length_device: Length of the device.

width_device: Width of the device.

Returns a shape function and a matrix of points representing the updated shape function.

ImagPhotonics.visualize_cell — Function

visualize_cell(cell [,transform])

Visualize the cell generated using the python package gdstk. The code only accepts object of type

gdstk.polygons

gdstk.paths

gdstk.references

The object gdstk.curve is not supported and it has to be converted into a polygon first

...

Arguments

cell::PyObject: the library object containing all the design information

transform::Array{Real,2}: it is a 2x3 transformation matrix containg the translation information in the last column

...

Examples

julia> visualize_cell(layout)
julia> ϕ=π/2;
julia> dx=1; dy=0;
julia> transform= [cos(ϕ) -sin(ϕ) dx;sin(ϕ) cos(ϕ),dy]
2x3 Array{Float64,2}:
 0 -1  1
 1  0  0
julia> visualize_cell(layout,transform)

See also visualize_gds

ImagPhotonics.visualize_gds — Method

visualize_gds(layout)

Visualize the layout generated using the python package gdstk. Layout is a PyObject <gdstk.Library> and it contains an array of cells PyObject <gdstk.Cells>

Arguments

layout::PyObject: the library object containing all the design information

Examples

julia> visualize_gds(layout)

See also visualize_cell